refactor: pickedNumber 클래스 분리
