+++ src/test/java/lotto/domain/BonusNumberTest.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public class BonusNumberTest {
+}
